+++ yast/requests.py
@@ -108,9 +108,8 @@ class Request(Mapping):
     def url_for(self, name: str, **path_params: typing.Any) -> URL:
         router = self._scope["router"]
 
-        # todo: routing
         url = router.url_path_for(name, **path_params)
-        return url.replace(secure=self.url.is_secure, netloc=self.url.netloc)
+        return url.make_absolute_url(base_url=self.url)
 
     async def stream(self):
         if hasattr(self, "_body"):